+++ osidb/__init__.py
@@ -2,5 +2,5 @@
 osidb version
 """
 
-__version__ = "5.0.1"
+__version__ = "5.1.0"
 VERSION = __version__