+++ docs/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+- Fix regressed query parameter types in v1 readonly APIs (OSIDB-4611)
+- Fix incorrect query parameters with "tracker" relationship in v1 readonly APIs (OSIDB-4611)
+- Fix include/exclude fields usage on v1 readyonl APIs (OSIDB-4611)
 
 ## [5.0.1] - 2025-11-03
 ### Fixed